+++ apps/admin/src/views/home/index.ts
@@ -1,0 +1,1 @@
+export { default as HomePage } from './ui/HomePage';
