+++ apps/admin/src/views/signin/index.ts
@@ -1,0 +1,1 @@
+export { default as SigninPage } from './ui/SigninPage';
