+++ apps/admin/src/views/signin/ui/SigninPage/index.tsx
@@ -1,15 +1,18 @@
 'use client';
 
+import { cn } from '@repo/shared/lib';
 import { SigninForm } from '@repo/shared/ui';
 
 const SigninPage = () => {
   return (
-    <SigninForm
-      title="Data GSM Admin"
-      onGoogleLogin={() => {
-        console.log('admin: google login click');
-      }}
-    />
+    <div className={cn('bg-background flex min-h-screen items-center justify-center px-4')}>
+      <SigninForm
+        title="Data GSM Admin"
+        onGoogleLogin={() => {
+          console.log('admin: google login click');
+        }}
+      />
+    </div>
   );
 };
